+++ craft_parts/overlays/overlay_manager.py
@@ -23,6 +23,8 @@ from collections.abc import Callable
 from pathlib import Path
 from typing import Any, Literal, TypeVar, cast
 
+from typing_extensions import Self
+
 from craft_parts import packages
 from craft_parts.infos import ProjectInfo
 from craft_parts.parts import Part
@@ -190,12 +192,7 @@ class OverlayManager:
 
         :param package_names: The list of packages to download.
         """
-        if not self._overlay_fs:
-            raise RuntimeError("overlay filesystem not mounted")
-
-        mount_dir = self._project_info.overlay_mount_dir
-        chroot.chroot(
-            mount_dir,
+        self.run(
             _defer_evaluation(packages.Repository.download_packages),
             package_names,
         )
@@ -205,15 +202,19 @@ class OverlayManager:
 
         :param package_names: The list of packages to install.
         """
-        if not self._overlay_fs:
-            raise RuntimeError("overlay filesystem not mounted")
-
-        mount_dir = self._project_info.overlay_mount_dir
-        chroot.chroot(
-            mount_dir,
+        self.run(
             _defer_evaluation(packages.Repository.install_packages),
             package_names,
             refresh_package_cache=False,
+        )
+
+    def run(self, target: Callable[..., _T], *args: Any, **kwargs: Any) -> _T:
+        """Run the given callable inside the chroot environment."""
+        if not self._overlay_fs:
+            raise RuntimeError("overlay filesystem not mounted")
+
+        return chroot.chroot(
+            self._project_info.overlay_mount_dir, target, *args, **kwargs
         )
 
 
@@ -237,7 +238,7 @@ class LayerMount:
         self._pkg_cache = pkg_cache
         self._pid = os.getpid()
 
-    def __enter__(self) -> "LayerMount":
+    def __enter__(self) -> Self:
         logger.debug("---- Enter layer mount context ----")
         self._overlay_manager.mount_layer(
             self._top_part,
@@ -272,7 +273,7 @@ class PackageCacheMount:
         self._overlay_manager.mkdirs()
         self._pid = os.getpid()
 
-    def __enter__(self) -> "PackageCacheMount":
+    def __enter__(self) -> Self:
         logger.debug("---- Enter package cache mount context ----")
         self._overlay_manager.mount_pkg_cache()
         return self
@@ -295,3 +296,11 @@ class PackageCacheMount:
         :param package_names: The list of packages to download.
         """
         self._overlay_manager.download_packages(package_names)
+
+
+class ChrootMount(LayerMount):
+    """Context manager that mounts an overlay for step processing and runs code inside a chroot environment."""
+
+    def __call__(self, target: Callable[..., _T], *args: Any, **kwargs: Any) -> _T:
+        """Synthax sugar method to run within chroot."""
+        return self._overlay_manager.run(target, *args, **kwargs)